Handle taskq creation failure in vdev_open_children()
When creating and destroying pools in tight loop it's possible to
exhaust the number of allowed threads on a system.  This results
in taskq_create() failling and a NULL dereference.

Resolve the issue by dispatching vdev_open_child() to the global
system taskq.  In the case where vdev_open_children() is called
from a system_taskq worker thread vdev_open_child() is processed
sequentially.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue zfsonlinux/spl#521