Routine to parse a rotor vector configuration. For testing, the comment pool property is used to store the config string. Config should be in its own property, at least. Perhaps even split into multiple properties, one per rotor, and limit separate from types.
Format:

RV:[[spec]<=limit]:spec

with spec being a comma-separated list of vdev-guids and generic
type specifiers: ssd, ssd-raidz, mixed, hdd, and hdd-raidz.

limit gives the limit in kbytes of which allocations are allowed within
the rotor vector.  The last rotor has no limit.  (And vdevs which are not
matched by the guids or the generic types are placed in the last rotor.)

For the moment things are a bit difficult due to the 32 character limit
of the comment field.  That is also why the limit is implicitly expressed in
units of 1024 bytes.

Example:

RV:ssd<=4:mixed<=64:123,hdd

Allocations less than 4 kbytes are allocated on an ssd-only vdev (mirror or not).
Allocations less than 64 kbytes end up on ssd/hdd mixed (mirrors, such raidz makes no sense).
Other allocations end up on remaining disks.
123 represents av vdev guid.