Report usage according to most full rotor category.
Adjust the dspace reported from metaslab_class such that we look full
when one rotor vector category (e.g. SSD or HDD) becomes full.  This
since large content for the HDD should not spill into the SSD (quickly
filling it up prematurely), and as we also do not want to spill small
content that should be on the SSD onto the HDD.